Rename index to run example outside dir
It can be used without editing:

    npm test

    NODE_PATH=. node test

    cd .. && NODE_PATH=. node bmp085-sensor/test
