+++ bmp085-sensor.js
